Split Logout component to have a container
